Adjust and fix duplicate code
