+++ src/auto_flutter_dig/module/aflutter/task/project/init/find/flavor/flavor.py
@@ -1,6 +1,6 @@
 from typing import Callable
 
-from ........core.utils import _Dict
+from ........core.utils.task.subtask_run_all import BaseSubtaskRunAll
 from ........model.task.group import TaskGroup
 from ........model.task.identity import TaskIdentity
 from ........model.task.init.project_identity import InitProjectTaskIdentity
@@ -13,7 +13,7 @@ from ........module.aflutter.task.project.init.find.flavor.web import ProjectIni
 from ........module.aflutter.task.project.init.find.platform import ProjectInitFindPlatformTask
 
 
-class ProjectInitFindFlavorIdentity(AflutterTaskIdentity, InitProjectTaskIdentity, TaskGroup):
+class _ProjectInitFindFlavorIdentity(AflutterTaskIdentity, InitProjectTaskIdentity, TaskGroup):
     def __init__(self, creator: Callable[[], Task]) -> None:
         InitProjectTaskIdentity.__init__(self, "", "", "", [], creator)
         AflutterTaskIdentity.__init__(self, "-project-init-find-flavor", "", [], creator)
@@ -32,20 +32,6 @@ class ProjectInitFindFlavorIdentity(AflutterTaskIdentity, InitProjectTaskIdentit
         return [ProjectInitFindPlatformTask.identity]
 
 
-class ProjectInitFindFlavorTask(Task):
-    identity: ProjectInitFindFlavorIdentity = ProjectInitFindFlavorIdentity(
-        lambda: ProjectInitFindFlavorTask(ProjectInitFindFlavorTask.identity)
-    )
-
-    def __init__(self, subtask: TaskGroup) -> None:
-        super().__init__()
-        self._subtask = subtask
-
-    def describe(self, args: Args) -> str:
-        return ""
-
-    def execute(self, args: Args) -> TaskResult:
-        tasks = _Dict.flatten(self._subtask.subtasks)
-        tasks.reverse()
-        self._append_task(tasks)
-        return TaskResult(args)
+ProjectInitFindFlavorIdentity: _ProjectInitFindFlavorIdentity = _ProjectInitFindFlavorIdentity(
+    lambda: BaseSubtaskRunAll(ProjectInitFindFlavorIdentity)
+)
