+++ src/auto_flutter_dig/module/firebase/task/setup/check.py
@@ -2,13 +2,13 @@ from typing import Optional
 
 from .....core.config import Config
 from .....core.string import SB
-from .....core.utils.task.process.timeout import BaseProcessTimeoutTask, Process, ProcessOrResult
+from .....core.utils.task.process.check import BaseProcessCheckTask, Process, ProcessOrResult
 from .....model.task.task import *  # pylint: disable=wildcard-import
 from .....module.firebase.identity import FirebaseTaskIdentity
 from .....module.firebase.model._const import *  # pylint: disable=wildcard-import
 
 
-class FirebaseCheck(BaseProcessTimeoutTask):
+class FirebaseCheck(BaseProcessCheckTask):
     identity = FirebaseTaskIdentity(
         "-firebase-check",
         "Checking firebase-cli",
@@ -25,21 +25,6 @@ class FirebaseCheck(BaseProcessTimeoutTask):
             arguments=[FIREBASE_DISABLE_INTERACTIVE_MODE.value, "--version"],
             environment=FIREBASE_ENV.value,
         )
-
-    def _on_interval(self, process: Process, elapsed: float, count: int) -> None:
-        if count == 1:
-            self._print("  Skill wating...")
-        elif count == 3:
-            self._print(SB().append("  It is taking some time...", SB.Color.YELLOW).str())
-        return super()._on_interval(process, elapsed, count)
-
-    def _on_process_stop(self, process: Process, elapsed: float, count: int) -> None:
-        self._print(SB().append("  Stop process...", SB.Color.RED).str())
-        return super()._on_process_stop(process, elapsed, count)
-
-    def _on_process_kill(self, process: Process, elapsed: float, count: int) -> None:
-        self._print(SB().append("  Kill process...", SB.Color.RED, True).str())
-        return super()._on_process_kill(process, elapsed, count)
 
     def _handle_process_exception(
         self,
